remove loss fuctions in metric.py and move loss test functions to test_loss.py
